+++ src/utils.ts
@@ -1,16 +1,16 @@
-import { dateTime, dateTimeAsMoment } from "@grafana/data";
-import { Option, ShiftData, ShiftI, TExtendedShift, TMappings, TUpdateActiveShiftProps } from "./types";
+import { dateTime, dateTimeAsMoment } from '@grafana/data';
+import { Option, ShiftData, ShiftI, TExtendedShift, TMappings, TUpdateActiveShiftProps } from './types';
 
 export const dateTimeFormat = `YYYY-MM-DD HH:mm:ss`;
-export const fakeEpoc = '2009-10-17'
+export const fakeEpoc = '2009-10-17';
 
 type TParsedShiftData = {
-  shiftGroupName: string
-  shiftGroupUUID: string
-  start: string
-  end: string
-  order: number
-}
+  shiftGroupName: string;
+  shiftGroupUUID: string;
+  start: string;
+  end: string;
+  order: number;
+};
 
 export const parseShiftData = (uuid: string, values: any[]): TParsedShiftData | null => {
   const { text: data } = values.find(({ value }) => value === uuid) || {};
@@ -30,19 +30,26 @@ export const parseShiftData = (uuid: string, values: any[]): TParsedShiftData | 
   };
 };
 
-export const transformShiftData = (shift: ShiftI & { index: number }, hasNextDayShift: boolean, productionDate: number): ShiftI & TExtendedShift => {
-  const currentHour = dateTimeAsMoment(productionDate).get('hour')
-  let startDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.start}`)
-  let endDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.end}`)
-  const isStartHourGreater = startHourIsGreater(shift.start, shift.end)
-  const isActive = isCurrentTimeInShiftRange(shift)
-  const endHour = endDate.get('hour')
+export const transformShiftData = (
+  shift: ShiftI & { index: number },
+  hasNextDayShift: boolean,
+  productionDate: number
+): ShiftI & TExtendedShift => {
+  const currentHour = dateTimeAsMoment(productionDate).get('hour');
+  let startDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.start}`);
+  let endDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.end}`);
+  const isStartHourGreater = startHourIsGreater(shift.start, shift.end);
+  const isActive = isCurrentTimeInShiftRange(shift);
+  const endHour = endDate.get('hour');
 
   if (hasNextDayShift && isStartHourGreater) {
     if (currentHour <= 23 && currentHour > endHour) {
-      endDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.end}`).add(1, 'day')
+      endDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.end}`).add(1, 'day');
     } else {
-      startDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.start}`).subtract(1, 'day')
+      startDate = dateTimeAsMoment(`${dateTime(productionDate).format('YYYY-MM-DD')} ${shift.start}`).subtract(
+        1,
+        'day'
+      );
     }
   }
 
@@ -50,62 +57,61 @@ export const transformShiftData = (shift: ShiftI & { index: number }, hasNextDay
     startDate: startDate.format(dateTimeFormat),
     endDate: endDate.format(dateTimeFormat),
     isActive,
-    timeStamps: [startDate.unix() * 1000, endDate.unix() * 1000]
-  }
+    timeStamps: [startDate.unix() * 1000, endDate.unix() * 1000],
+  };
 
   return {
     ...shift,
     startDate,
     endDate,
     _,
-  }
-}
+  };
+};
 
 export const startHourIsGreater = (startHour: string, endHour: string) => {
-  const startDateHour = dateTimeAsMoment(`${dateTime().format('YYYY-MM-DD')} ${startHour}`).get('hour')
-  const endDateHour = dateTimeAsMoment(`${dateTime().format('YYYY-MM-DD')} ${endHour}`).get('hour')
+  const startDateHour = dateTimeAsMoment(`${dateTime().format('YYYY-MM-DD')} ${startHour}`).get('hour');
+  const endDateHour = dateTimeAsMoment(`${dateTime().format('YYYY-MM-DD')} ${endHour}`).get('hour');
 
-  return startDateHour > endDateHour
-}
+  return startDateHour > endDateHour;
+};
 
 export const getHasNextDayShift = (options: Option[], optionsData: any[]) => {
   return options.reduce((res: boolean, option) => {
     if (res) {
-      return true
+      return true;
     }
 
     const data = parseShiftData(option.value, optionsData);
 
     if (!data) {
-      return false
+      return false;
     }
 
-    return startHourIsGreater(data.start, data.end)
-
+    return startHourIsGreater(data.start, data.end);
   }, false);
-}
+};
 
 export const isCurrentTimeInShiftRange = (shift: ShiftI) => {
-  let startDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.start}`).unix()
-  let endDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.end}`).unix()
-  const currentTime = dateTimeAsMoment(`${fakeEpoc} ${dateTimeAsMoment().format('HH:mm:ss')}`).unix()
-  const nowHour = dateTimeAsMoment().get('hour')
+  let startDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.start}`).unix();
+  let endDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.end}`).unix();
+  const currentTime = dateTimeAsMoment(`${fakeEpoc} ${dateTimeAsMoment().format('HH:mm:ss')}`).unix();
+  const nowHour = dateTimeAsMoment().get('hour');
 
   if (startHourIsGreater(shift.start, shift.end)) {
     if (nowHour <= 23 && currentTime > dateTimeAsMoment(`${fakeEpoc} ${shift.start}`).unix()) {
-      endDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.end}`).add(1, 'day').unix()
+      endDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.end}`).add(1, 'day').unix();
     } else {
-      startDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.start}`).subtract(1, 'day').unix()
+      startDateTime = dateTimeAsMoment(`${fakeEpoc} ${shift.start}`).subtract(1, 'day').unix();
     }
   }
 
-  const state = currentTime >= startDateTime && currentTime < endDateTime
+  const state = currentTime >= startDateTime && currentTime < endDateTime;
 
-  return state
-}
+  return state;
+};
 
 export const getShifts = (options: Option[], optionsData: any[], productionDate: number) => {
-  const hasNextDayShift: boolean = getHasNextDayShift(options, optionsData)
+  const hasNextDayShift: boolean = getHasNextDayShift(options, optionsData);
   const shifts: ShiftData = options.reduce((res: any, { text, value: uuid }, index: number) => {
     const data = parseShiftData(uuid, optionsData);
 
@@ -117,12 +123,16 @@ export const getShifts = (options: Option[], optionsData: any[], productionDate:
       ...res,
       [data.shiftGroupUUID]: [
         ...(res[data.shiftGroupUUID] || []),
-        transformShiftData({
-          ...data,
-          uuid,
-          label: text,
-          index,
-        }, hasNextDayShift, productionDate),
+        transformShiftData(
+          {
+            ...data,
+            uuid,
+            label: text,
+            index,
+          },
+          hasNextDayShift,
+          productionDate
+        ),
       ],
     };
   }, {});
@@ -184,16 +194,18 @@ export function getRelativeDates() {
 
 export const updateActiveShift = (props: TUpdateActiveShiftProps) => {
   const isRealtimeActive = !!(props.isAutoSelectShift && props.autoSelectShiftGroup);
-  const relativeToDate = isRealtimeActive ? dateTimeAsMoment().unix() * 1000 : props.productionDate
+  const relativeToDate = isRealtimeActive ? dateTimeAsMoment().unix() * 1000 : props.productionDate;
   const shifts = getShifts(props.shifts.options, props.shifts.values, relativeToDate);
 
   const [initGroup] = Object.keys(shifts) || [];
   let activeShifts =
-    props.autoSelectShiftGroup && shifts[props.autoSelectShiftGroup] ? shifts[props.autoSelectShiftGroup] : shifts[initGroup];
+    props.autoSelectShiftGroup && shifts[props.autoSelectShiftGroup]
+      ? shifts[props.autoSelectShiftGroup]
+      : shifts[initGroup];
 
-  const activeShift = (activeShifts as unknown as TExtendedShift[]).find(({ _ }) => _.isActive)
-  console.log('SHIFTS', activeShift)
-  console.log('isRealtimeActive', isRealtimeActive)
+  const activeShift = ((activeShifts as unknown) as TExtendedShift[]).find(({ _ }) => _.isActive);
+  console.log('SHIFTS', activeShift);
+  console.log('isRealtimeActive', isRealtimeActive);
 
   if (activeShift) {
     props.setShiftParams(activeShift, false);
@@ -273,4 +285,4 @@ export const updateActiveShift = (props: TUpdateActiveShiftProps) => {
 
   //     return item;
   //   });
-}
+};