+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.0
+
+- Realtime bugs fixed
+- Localisation for labels added
+- Refactored
+- Panel properties optimised
+
 ## 0.0.3
 
 Realtime Shift feature and, extra panel options added to customize the panel. Like changing labels and display types.